alteracao de validacoes
